chore: raise PHPStan to level 9 + adopt phpstan-strict-rules
Level 9 bans operations on `mixed` without narrowing; all errors cleared with
behavior-preserving narrowing. Also adopt phpstan-strict-rules (dev dep) — the
full set minus the WordPress/node-runtime idiom-fighters (empty(), truthy
conditionals, short ternary, noVariableVariables (Node $this->{$name} per
make_node decision #11), checkDynamicProperties) — adding redundant-cast
removal, explicit array_filter callbacks, ==/===, and parent::__construct()
calls on the Node subclasses (base Node::__construct() is idempotent, so safe).

A full multi-agent review confirmed no behavior regressions (the command
rate-limit $burst stays int-typed; the existing clamp normalizes it).

Behavior-preserving: full PHPUnit suite green (1732).

+++ includes/class-partition-node.php
@@ -416,7 +416,7 @@ class Partition_Node extends Timer_Node {
 			$this->heartbeat_timer->name( "{$this->name}:heartbeat" );
 			$this->heartbeat_timer->sink( $this->write_lock );
 			$this->heartbeat_timer->set_key( 'heartbeat' );
-			$this->heartbeat_timer->set_timer( (int) ( $stale_timeout * 1000 / 3 ) );
+			$this->heartbeat_timer->set_timer( $stale_timeout * 1000 / 3 );
 			$this->heartbeat_timer->patron( $this );
 		}
